Maxima and Minima of Random Variables

Returns the maxima and minima of the components of a random vector.

rvmin applies the function min to each component of the argument x. Missing values are removed.

rvmax applies the function max to each component of the argument x. Missing values are removed.

rvrange applies the function range to each component of the argument x. Missing values are removed.

min.rv returns the minimum of the random vector, returning thus one random variable. Similarly max.rv returns the maximum of a vector.

pmin.rv and pmax.rv returns the componentwise minima or maxima of several random vectors or constants, yielding thus a random vector of the same length.

Arguments

x
an rv or rvsummary object

na.rm
remove missing values?

…
one or more rv objects or numeric objects

Value

A numeric vector of the same dimension as x.

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

See also

Examples

  x <- rvpois(10, lambda=3)
  rvmin(x)

#>  [1] 0 0 0 0 0 0 0 0 0 0

  rvmax(x)

#>  [1] 11 10 11 11 12 10 12 10 10 10

  rvrange(x)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    0    0    0    0    0    0    0    0    0     0
#> [2,]   11   10   11   11   12   10   12   10   10    10