Mathematical functions and Operators for rv Objects

Mathematical functions and operators adapted to work with random variable (rv) objects.

# S3 method for rv
Math(x, ...)

# S3 method for rv
Ops(e1, e2 = NULL)

Arguments

x

object

...

further arguments passed to or from other methods

e1

object

e2

object

Details

The operator method preserves the names of the longer vector (or those of the first if the lengths match).

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Author

Jouni Kerman jouni@kerman.com

Examples

 
  x <- rvnorm(10)
  -x
#>         mean   sd   1% 2.5%   25%     50%  75% 97.5% 99% sims
#>  [1]  0.0079 1.01 -2.3 -2.0 -0.66  0.0186 0.68   2.0 2.3 4000
#>  [2] -0.0193 0.99 -2.3 -2.0 -0.70 -0.0335 0.65   2.0 2.3 4000
#>  [3]  0.0201 1.00 -2.2 -1.9 -0.67  0.0356 0.70   2.0 2.3 4000
#>  [4] -0.0017 1.01 -2.3 -2.0 -0.69 -0.0082 0.66   2.0 2.3 4000
#>  [5] -0.0184 1.01 -2.3 -2.0 -0.70 -0.0433 0.67   2.0 2.4 4000
#>  [6]  0.0109 1.00 -2.3 -1.9 -0.67  0.0057 0.70   2.0 2.3 4000
#>  [7] -0.0168 1.00 -2.3 -2.0 -0.67 -0.0112 0.64   2.0 2.3 4000
#>  [8]  0.0039 0.98 -2.2 -1.9 -0.68  0.0029 0.66   1.9 2.3 4000
#>  [9]  0.0039 1.01 -2.4 -1.9 -0.67  0.0067 0.69   2.0 2.3 4000
#> [10] -0.0185 1.02 -2.4 -2.0 -0.72 -0.0169 0.68   2.0 2.4 4000
  names(x) <- paste("x[", seq_along(x), "]", sep="")
  x + 1:10
#>       name  mean   sd    1%   2.5%  25%   50%  75% 97.5%  99% sims
#>  [1]  x[1]  0.99 1.01 -1.29 -1.002 0.32  0.98  1.7   3.0  3.3 4000
#>  [2]  x[2]  2.02 0.99 -0.31  0.023 1.35  2.03  2.7   4.0  4.3 4000
#>  [3]  x[3]  2.98 1.00  0.67  1.003 2.30  2.96  3.7   4.9  5.2 4000
#>  [4]  x[4]  4.00 1.01  1.68  2.048 3.34  4.01  4.7   6.0  6.3 4000
#>  [5]  x[5]  5.02 1.01  2.63  3.038 4.33  5.04  5.7   7.0  7.3 4000
#>  [6]  x[6]  5.99 1.00  3.71  4.028 5.30  5.99  6.7   7.9  8.3 4000
#>  [7]  x[7]  7.02 1.00  4.72  5.049 6.36  7.01  7.7   9.0  9.3 4000
#>  [8]  x[8]  8.00 0.98  5.67  6.051 7.34  8.00  8.7   9.9 10.2 4000
#>  [9]  x[9]  9.00 1.01  6.70  7.004 8.31  8.99  9.7  10.9 11.4 4000
#> [10] x[10] 10.02 1.02  7.63  7.982 9.32 10.02 10.7  12.0 12.4 4000
  1:2 + x
#>       name mean   sd    1%   2.5%  25%  50% 75% 97.5% 99% sims
#>  [1]  x[1] 0.99 1.01 -1.29 -1.002 0.32 0.98 1.7   3.0 3.3 4000
#>  [2]  x[2] 2.02 0.99 -0.31  0.023 1.35 2.03 2.7   4.0 4.3 4000
#>  [3]  x[3] 0.98 1.00 -1.33 -0.997 0.30 0.96 1.7   2.9 3.2 4000
#>  [4]  x[4] 2.00 1.01 -0.32  0.048 1.34 2.01 2.7   4.0 4.3 4000
#>  [5]  x[5] 1.02 1.01 -1.37 -0.962 0.33 1.04 1.7   3.0 3.3 4000
#>  [6]  x[6] 1.99 1.00 -0.29  0.028 1.30 1.99 2.7   3.9 4.3 4000
#>  [7]  x[7] 1.02 1.00 -1.28 -0.951 0.36 1.01 1.7   3.0 3.3 4000
#>  [8]  x[8] 2.00 0.98 -0.33  0.051 1.34 2.00 2.7   3.9 4.2 4000
#>  [9]  x[9] 1.00 1.01 -1.30 -0.996 0.31 0.99 1.7   2.9 3.4 4000
#> [10] x[10] 2.02 1.02 -0.37 -0.018 1.32 2.02 2.7   4.0 4.4 4000
  cumsum(x)
#>          mean  sd   1% 2.5%   25%     50%  75% 97.5% 99% sims
#>  [1] -0.00794 1.0 -2.3 -2.0 -0.68 -0.0186 0.66   2.0 2.3 4000
#>  [2]  0.01133 1.4 -3.3 -2.8 -0.95  0.0416 0.98   2.8 3.3 4000
#>  [3] -0.00872 1.7 -4.1 -3.4 -1.18  0.0041 1.14   3.4 3.9 4000
#>  [4] -0.00701 2.0 -4.8 -3.9 -1.37 -0.0133 1.35   4.0 4.8 4000
#>  [5]  0.01140 2.3 -5.4 -4.4 -1.48  0.0083 1.51   4.4 5.3 4000
#>  [6]  0.00048 2.5 -5.7 -4.8 -1.61 -0.0492 1.67   4.8 5.9 4000
#>  [7]  0.01724 2.6 -6.2 -5.1 -1.72 -0.0282 1.75   5.3 6.2 4000
#>  [8]  0.01338 2.9 -6.9 -5.7 -1.81 -0.0083 1.91   5.6 6.6 4000
#>  [9]  0.00943 3.0 -7.1 -6.0 -1.96 -0.0384 2.04   6.0 7.0 4000
#> [10]  0.02796 3.2 -7.5 -6.3 -2.13 -0.0290 2.18   6.3 7.4 4000
  cumprod(exp(x))
#>       mean     sd      1%   2.5%  25%  50% 75% 97.5%  99% sims
#>  [1]   1.7    2.1 0.10147 0.1351 0.51 0.98 1.9   7.2   10 4000
#>  [2]   2.7    5.7 0.03515 0.0600 0.39 1.04 2.7  15.7   27 4000
#>  [3]   4.5   17.4 0.01742 0.0328 0.31 1.00 3.1  30.2   51 4000
#>  [4]   8.5   80.3 0.00835 0.0199 0.25 0.99 3.8  55.3  118 4000
#>  [5]  15.3  173.2 0.00441 0.0118 0.23 1.01 4.5  81.7  209 4000
#>  [6]  25.6  391.3 0.00321 0.0084 0.20 0.95 5.3 125.2  347 4000
#>  [7]  35.6  409.6 0.00212 0.0059 0.18 0.97 5.8 191.2  498 4000
#>  [8]  46.5  400.2 0.00097 0.0033 0.16 0.99 6.7 281.4  758 4000
#>  [9]  68.0  750.0 0.00079 0.0025 0.14 0.96 7.7 385.3 1088 4000
#> [10] 111.8 1602.5 0.00057 0.0019 0.12 0.97 8.9 530.0 1648 4000