Add (Random) Straight Lines to a Plot

abline.rv, with random arguments (i.e. arguments of which at least one is an rv object), plots a sample of lines corresponding to of simulations of rv object x. If the arguments are all numeric (none is an rv object), the function call is passed on to abline.

abline.rv(a = NULL, b = NULL, h = NULL, v = NULL, ...)

Arguments

a

intercept

b

slope

h

y-value(s) horizontal line(s)

v

x-value(s) horizontal line(s)

...

further arguments passed to abline

Details

This is a version of abline that accepts random variable objects for the arguments a, b, h, or v.

The number of lines is determined by rvpar("line.sample"), default 20.

See the original help page in package `graphics.'

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Author

Jouni Kerman jouni@kerman.com

Examples

  if (FALSE) {
     demo("rvexample1")
  }