Distribution of the Range of a Random Vector

range.rv returns a 2-component random vector containing the distributions of the minimum and the maximum values of all the given arguments.

# S3 method for rv
range(..., na.rm = FALSE, finite = FALSE)

Arguments

...
further arguments passed to or from other methods

na.rm
logical, indicating if NAs should be omitted

finite
logical, indicating if all non-finite elements should be omitted

Details

This is the rv-compatible version of the function range.

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

See also

Examples

  x <- rvnorm(mean=1:10, sd=1)
  print(range(x))

#>     name  mean   sd   1%  2.5%  25%   50%  75% 97.5%  99% sims
#> [1]  min  0.74 0.89 -1.4 -0.87 0.09  0.84  1.3   2.2  2.4  200
#> [2]  max 10.40 0.84  8.7  9.05 9.82 10.30 10.9  12.2 12.4  200

  print(quantile(x, c(0,1)))

#>     name  mean   sd   1%  2.5%  25%   50%  75% 97.5%  99% sims
#> [1]   0%  0.74 0.89 -1.4 -0.87 0.09  0.84  1.3   2.2  2.4  200
#> [2] 100% 10.40 0.84  8.7  9.05 9.82 10.30 10.9  12.2 12.4  200