Generate Random Vectors from a Beta Sampling Model

rvbeta generates a random vector from the beta sampling model;

rvbeta(n = 1, shape1, shape2)

Arguments

n
integer, number of random variables to generate

shape1
positive number or rv, 1st shape parameter

shape2
positive number or rv, 2nd shape parameter

Details

rvnbeta(n, a, b) ("neutral" Beta distribution) is equivalent to rvbeta(n, 1/3+a, 1/3+b).

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Examples

   n <- 12         # sample size
   y <- (0:(n-1))  # observations
   a <- b <- 1/3   # the neutral beta prior
   rvbeta(1, shape1=a+y, shape2=b+n-y)

#>       mean    sd      1%    2.5%    25%    50%   75% 97.5%  99% sims
#>  [1] 0.027 0.044 1.1e-06 1.6e-06 0.0012 0.0077 0.037  0.14 0.24  200
#>  [2] 0.101 0.081 2.7e-03 8.3e-03 0.0404 0.0790 0.141  0.29 0.39  200
#>  [3] 0.183 0.092 1.8e-02 4.3e-02 0.1212 0.1737 0.233  0.40 0.46  200
#>  [4] 0.256 0.118 4.5e-02 6.2e-02 0.1730 0.2373 0.325  0.51 0.58  200
#>  [5] 0.353 0.127 1.0e-01 1.4e-01 0.2525 0.3497 0.434  0.61 0.62  200
#>  [6] 0.419 0.132 1.7e-01 2.1e-01 0.3213 0.4143 0.506  0.70 0.73  200
#>  [7] 0.505 0.139 2.1e-01 2.4e-01 0.4014 0.5025 0.599  0.79 0.80  200
#>  [8] 0.576 0.145 2.2e-01 2.8e-01 0.4671 0.5890 0.682  0.85 0.87  200
#>  [9] 0.665 0.137 3.6e-01 3.9e-01 0.5692 0.6915 0.764  0.88 0.90  200
#> [10] 0.733 0.114 4.6e-01 4.9e-01 0.6565 0.7315 0.818  0.93 0.95  200
#> [11] 0.805 0.124 4.7e-01 4.9e-01 0.7409 0.8272 0.903  0.97 0.97  200
#> [12] 0.895 0.084 6.5e-01 6.8e-01 0.8501 0.9114 0.963  0.99 1.00  200

   rvnbeta(1, shape1=y, shape2=n-y)

#>       mean    sd      1%    2.5%     25%    50%   75% 97.5%  99% sims
#>  [1] 0.024 0.039 2.5e-08 4.2e-07 0.00065 0.0064 0.029  0.16 0.18  200
#>  [2] 0.096 0.074 3.8e-03 6.2e-03 0.03792 0.0765 0.131  0.26 0.27  200
#>  [3] 0.172 0.100 2.1e-02 2.6e-02 0.10308 0.1482 0.233  0.40 0.46  200
#>  [4] 0.273 0.126 5.0e-02 6.5e-02 0.18405 0.2493 0.360  0.53 0.61  200
#>  [5] 0.349 0.132 1.2e-01 1.4e-01 0.24441 0.3239 0.445  0.60 0.67  200
#>  [6] 0.420 0.131 1.4e-01 1.6e-01 0.32510 0.4188 0.515  0.67 0.74  200
#>  [7] 0.486 0.132 2.2e-01 2.5e-01 0.39078 0.4827 0.575  0.72 0.77  200
#>  [8] 0.578 0.144 2.9e-01 3.0e-01 0.47713 0.5809 0.685  0.83 0.89  200
#>  [9] 0.654 0.131 3.1e-01 3.8e-01 0.56734 0.6666 0.761  0.86 0.89  200
#> [10] 0.738 0.119 4.5e-01 4.8e-01 0.66104 0.7533 0.823  0.93 0.94  200
#> [11] 0.812 0.098 5.4e-01 5.8e-01 0.75694 0.8260 0.878  0.97 0.98  200
#> [12] 0.896 0.085 6.2e-01 6.4e-01 0.85739 0.9151 0.957  0.99 1.00  200