Credible (Uncertainty) Intervals for Random Scalars

Computes credible (uncertainty) intervals for a given vector, given quantiles or the size of the middle interval

rvci(obj, interval = 0.95, one.sided = FALSE, left = TRUE)

Arguments

obj
random scalar or vector

interval
size of the middle interval or the quantile range of the interval

one.sided
logical, FALSE if two-sided interval is desired

left
logical, indicating if the left one-sided interval is desired

Value

For two-sided intervals, an array of numbers of dimension c(2,length(x)), for one-sided intervals, a vector of the same length as x.

Details

If interval is of length two or more, the return value will be the quantiles given by range(interval).

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Examples

  rvci(rvnorm(1), interval=0.683) # Should be about c(-1,1).

#>        15.85%    84.15%
#> [1,] -1.18051 0.8543461