Histogram of Distributions of Components of a Random Vector

rvhist shows a grid of histograms of simulations of the components of a random vector.

rvhist(x, ...)

Arguments

x
an rv object

…
further arguments passed to the function hist

Details

Outputs a histogram using the hist function with the option freq=FALSE. This can be overridden by specifying the argument freq or prob. See the function hist for details.

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").