Mathematical functions and Operators for rv Objects

Mathematical functions and operators adapted to work with random variable (rv) objects.

# S3 method for rv
Math(x, ...)

# S3 method for rv
Ops(e1, e2 = NULL)

Arguments

x
object

...
further arguments passed to or from other methods

e1
object

e2
object

Details

The operator method preserves the names of the longer vector (or those of the first if the lengths match).

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Examples

  x <- rvnorm(10)
  -x

#>         mean   sd   1% 2.5%   25%     50%  75% 97.5% 99% sims
#>  [1]  0.0146 1.00 -2.3 -1.9 -0.67  0.0302 0.69   1.9 2.2 4000
#>  [2]  0.0099 0.99 -2.2 -1.9 -0.68 -0.0190 0.70   1.9 2.3 4000
#>  [3] -0.0074 1.00 -2.4 -2.0 -0.68 -0.0095 0.66   2.0 2.2 4000
#>  [4]  0.0226 1.00 -2.3 -2.0 -0.64  0.0492 0.69   1.9 2.2 4000
#>  [5]  0.0078 0.99 -2.3 -2.0 -0.65  0.0178 0.67   1.9 2.3 4000
#>  [6] -0.0038 1.02 -2.3 -1.9 -0.70 -0.0049 0.70   2.0 2.4 4000
#>  [7] -0.0051 0.99 -2.3 -1.9 -0.68  0.0236 0.65   1.9 2.3 4000
#>  [8] -0.0139 1.01 -2.3 -1.9 -0.74 -0.0260 0.67   2.0 2.3 4000
#>  [9] -0.0108 0.99 -2.3 -2.0 -0.66 -0.0118 0.67   1.9 2.3 4000
#> [10] -0.0065 0.99 -2.3 -1.9 -0.67 -0.0115 0.65   1.9 2.2 4000

  names(x) <- paste("x[", seq_along(x), "]", sep="")
  x + 1:10

#>       name  mean   sd    1%  2.5%  25%   50%  75% 97.5%  99% sims
#>  [1]  x[1]  0.99 1.00 -1.24 -0.93 0.31  0.97  1.7   2.9  3.3 4000
#>  [2]  x[2]  1.99 0.99 -0.29  0.08 1.30  2.02  2.7   3.9  4.2 4000
#>  [3]  x[3]  3.01 1.00  0.76  1.04 2.34  3.01  3.7   5.0  5.4 4000
#>  [4]  x[4]  3.98 1.00  1.78  2.07 3.31  3.95  4.6   6.0  6.3 4000
#>  [5]  x[5]  4.99 0.99  2.70  3.08 4.33  4.98  5.7   7.0  7.3 4000
#>  [6]  x[6]  6.00 1.02  3.62  4.02 5.30  6.00  6.7   7.9  8.3 4000
#>  [7]  x[7]  7.01 0.99  4.70  5.06 6.35  6.98  7.7   8.9  9.3 4000
#>  [8]  x[8]  8.01 1.01  5.71  6.04 7.33  8.03  8.7   9.9 10.3 4000
#>  [9]  x[9]  9.01 0.99  6.72  7.09 8.33  9.01  9.7  11.0 11.3 4000
#> [10] x[10] 10.01 0.99  7.76  8.12 9.35 10.01 10.7  11.9 12.3 4000

  1:2 + x

#>       name mean   sd    1%   2.5%  25%  50% 75% 97.5% 99% sims
#>  [1]  x[1] 0.99 1.00 -1.24 -0.931 0.31 0.97 1.7   2.9 3.3 4000
#>  [2]  x[2] 1.99 0.99 -0.29  0.080 1.30 2.02 2.7   3.9 4.2 4000
#>  [3]  x[3] 1.01 1.00 -1.24 -0.965 0.34 1.01 1.7   3.0 3.4 4000
#>  [4]  x[4] 1.98 1.00 -0.22  0.073 1.31 1.95 2.6   4.0 4.3 4000
#>  [5]  x[5] 0.99 0.99 -1.30 -0.921 0.33 0.98 1.7   3.0 3.3 4000
#>  [6]  x[6] 2.00 1.02 -0.38  0.017 1.30 2.00 2.7   3.9 4.3 4000
#>  [7]  x[7] 1.01 0.99 -1.30 -0.941 0.35 0.98 1.7   2.9 3.3 4000
#>  [8]  x[8] 2.01 1.01 -0.29  0.041 1.33 2.03 2.7   3.9 4.3 4000
#>  [9]  x[9] 1.01 0.99 -1.28 -0.907 0.33 1.01 1.7   3.0 3.3 4000
#> [10] x[10] 2.01 0.99 -0.24  0.115 1.35 2.01 2.7   3.9 4.3 4000

  cumsum(x)

#>         mean  sd   1% 2.5%   25%     50%  75% 97.5% 99% sims
#>  [1] -0.0146 1.0 -2.2 -1.9 -0.69 -0.0302 0.67   1.9 2.3 4000
#>  [2] -0.0245 1.4 -3.2 -2.7 -0.97 -0.0320 0.94   2.7 3.1 4000
#>  [3] -0.0171 1.7 -4.0 -3.4 -1.16 -0.0334 1.15   3.4 3.9 4000
#>  [4] -0.0397 2.0 -4.6 -3.9 -1.38 -0.0826 1.30   3.9 4.6 4000
#>  [5] -0.0475 2.2 -5.4 -4.4 -1.58 -0.0834 1.48   4.4 5.1 4000
#>  [6] -0.0437 2.5 -5.7 -4.9 -1.64 -0.0582 1.65   4.8 5.8 4000
#>  [7] -0.0385 2.7 -6.4 -5.3 -1.76 -0.0286 1.79   5.2 6.2 4000
#>  [8] -0.0246 2.8 -6.6 -5.6 -1.93 -0.0210 1.86   5.4 6.6 4000
#>  [9] -0.0139 3.0 -6.9 -5.8 -2.03  0.0082 1.97   5.7 6.9 4000
#> [10] -0.0074 3.1 -7.1 -6.0 -2.15  0.0021 2.07   6.0 7.4 4000

  cumprod(exp(x))

#>       mean     sd      1%   2.5%  25%  50% 75% 97.5%    99% sims
#>  [1]   1.6    2.1 0.10594 0.1449 0.50 0.97 1.9   6.9    9.6 4000
#>  [2]   2.5    5.5 0.03929 0.0677 0.38 0.97 2.6  14.4   22.5 4000
#>  [3]   4.2   14.2 0.01815 0.0350 0.31 0.97 3.2  29.2   51.2 4000
#>  [4]   6.9   31.7 0.01005 0.0193 0.25 0.92 3.7  50.1  100.1 4000
#>  [5]  11.0   63.3 0.00462 0.0124 0.21 0.92 4.4  79.7  158.0 4000
#>  [6]  17.3  108.4 0.00319 0.0073 0.19 0.94 5.2 127.4  321.5 4000
#>  [7]  27.2  237.3 0.00164 0.0048 0.17 0.97 6.0 174.3  474.1 4000
#>  [8]  54.5  914.5 0.00130 0.0037 0.14 0.98 6.4 227.6  732.6 4000
#>  [9]  75.0  946.3 0.00099 0.0030 0.13 1.01 7.2 304.6  996.1 4000
#> [10] 123.8 1676.2 0.00082 0.0025 0.12 1.00 7.9 409.6 1643.9 4000