Mathematical functions and Operators for rv Objects

Mathematical functions and operators adapted to work with random variable (rv) objects.

# S3 method for rv
Math(x, ...)

# S3 method for rv
Ops(e1, e2 = NULL)

Arguments

x

object

...

further arguments passed to or from other methods

e1

object

e2

object

Details

The operator method preserves the names of the longer vector (or those of the first if the lengths match).

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Author

Jouni Kerman jouni@kerman.com

Examples

 
  x <- rvnorm(10)
  -x
#>         mean   sd   1% 2.5%   25%     50%  75% 97.5% 99% sims
#>  [1]  0.0144 1.00 -2.3 -1.9 -0.66  0.0223 0.69   2.0 2.3 4000
#>  [2]  0.0049 1.02 -2.4 -2.0 -0.69 -0.0133 0.69   2.0 2.3 4000
#>  [3] -0.0144 0.98 -2.4 -1.9 -0.68 -0.0105 0.65   1.9 2.2 4000
#>  [4]  0.0108 1.01 -2.5 -2.0 -0.65  0.0233 0.70   2.0 2.3 4000
#>  [5]  0.0295 1.00 -2.2 -1.9 -0.68  0.0278 0.70   2.0 2.3 4000
#>  [6]  0.0069 1.00 -2.3 -2.0 -0.68  0.0111 0.68   2.0 2.4 4000
#>  [7]  0.0172 0.99 -2.3 -1.9 -0.67  0.0258 0.71   1.9 2.3 4000
#>  [8]  0.0013 1.02 -2.4 -2.0 -0.67  0.0005 0.69   2.0 2.3 4000
#>  [9]  0.0063 1.01 -2.3 -2.0 -0.66 -0.0059 0.70   2.0 2.3 4000
#> [10] -0.0188 1.00 -2.4 -2.0 -0.68 -0.0104 0.65   1.9 2.2 4000
  names(x) <- paste("x[", seq_along(x), "]", sep="")
  x + 1:10
#>       name  mean   sd    1%   2.5%  25%   50%  75% 97.5%  99% sims
#>  [1]  x[1]  0.99 1.00 -1.33 -0.988 0.31  0.98  1.7   2.9  3.3 4000
#>  [2]  x[2]  2.00 1.02 -0.30  0.016 1.31  2.01  2.7   4.0  4.4 4000
#>  [3]  x[3]  3.01 0.98  0.81  1.106 2.35  3.01  3.7   4.9  5.4 4000
#>  [4]  x[4]  3.99 1.01  1.66  2.049 3.30  3.98  4.6   6.0  6.5 4000
#>  [5]  x[5]  4.97 1.00  2.66  2.951 4.30  4.97  5.7   6.9  7.2 4000
#>  [6]  x[6]  5.99 1.00  3.60  4.010 5.32  5.99  6.7   8.0  8.3 4000
#>  [7]  x[7]  6.98 0.99  4.72  5.050 6.29  6.97  7.7   8.9  9.3 4000
#>  [8]  x[8]  8.00 1.02  5.72  6.045 7.31  8.00  8.7  10.0 10.4 4000
#>  [9]  x[9]  8.99 1.01  6.67  7.046 8.30  9.01  9.7  11.0 11.3 4000
#> [10] x[10] 10.02 1.00  7.76  8.067 9.35 10.01 10.7  12.0 12.4 4000
  1:2 + x
#>       name mean   sd    1%   2.5%  25%  50% 75% 97.5% 99% sims
#>  [1]  x[1] 0.99 1.00 -1.33 -0.988 0.31 0.98 1.7   2.9 3.3 4000
#>  [2]  x[2] 2.00 1.02 -0.30  0.016 1.31 2.01 2.7   4.0 4.4 4000
#>  [3]  x[3] 1.01 0.98 -1.19 -0.894 0.35 1.01 1.7   2.9 3.4 4000
#>  [4]  x[4] 1.99 1.01 -0.34  0.049 1.30 1.98 2.6   4.0 4.5 4000
#>  [5]  x[5] 0.97 1.00 -1.34 -1.049 0.30 0.97 1.7   2.9 3.2 4000
#>  [6]  x[6] 1.99 1.00 -0.40  0.010 1.32 1.99 2.7   4.0 4.3 4000
#>  [7]  x[7] 0.98 0.99 -1.28 -0.950 0.29 0.97 1.7   2.9 3.3 4000
#>  [8]  x[8] 2.00 1.02 -0.28  0.045 1.31 2.00 2.7   4.0 4.4 4000
#>  [9]  x[9] 0.99 1.01 -1.33 -0.954 0.30 1.01 1.7   3.0 3.3 4000
#> [10] x[10] 2.02 1.00 -0.24  0.067 1.35 2.01 2.7   4.0 4.4 4000
  cumsum(x)
#>         mean  sd   1% 2.5%   25%      50%  75% 97.5% 99% sims
#>  [1] -0.0144 1.0 -2.3 -2.0 -0.69 -0.02235 0.66   1.9 2.3 4000
#>  [2] -0.0193 1.4 -3.2 -2.8 -0.99 -0.02245 0.97   2.7 3.2 4000
#>  [3] -0.0049 1.8 -4.0 -3.5 -1.18 -0.00539 1.20   3.4 4.0 4000
#>  [4] -0.0158 2.0 -4.7 -4.0 -1.42  0.00063 1.36   3.9 4.7 4000
#>  [5] -0.0453 2.2 -5.3 -4.3 -1.56 -0.05603 1.48   4.4 5.3 4000
#>  [6] -0.0521 2.4 -5.6 -4.8 -1.67 -0.04591 1.58   4.6 5.8 4000
#>  [7] -0.0694 2.6 -6.2 -5.1 -1.81 -0.11315 1.65   5.1 6.2 4000
#>  [8] -0.0707 2.8 -6.8 -5.5 -1.99 -0.08957 1.79   5.5 6.8 4000
#>  [9] -0.0770 3.0 -7.2 -5.9 -2.10 -0.10004 1.90   6.0 7.0 4000
#> [10] -0.0582 3.2 -7.5 -6.3 -2.15 -0.11992 2.09   6.2 7.6 4000
  cumprod(exp(x))
#>       mean     sd      1%   2.5%  25%  50% 75% 97.5%  99% sims
#>  [1]   1.6    2.2 0.09766 0.1370 0.50 0.98 1.9   6.7   10 4000
#>  [2]   2.7    6.8 0.03911 0.0588 0.37 0.98 2.7  15.4   24 4000
#>  [3]   4.2   11.9 0.01816 0.0300 0.31 0.99 3.3  29.0   57 4000
#>  [4]   7.4   35.4 0.00902 0.0178 0.24 1.00 3.9  49.6  110 4000
#>  [5]  12.7   95.3 0.00489 0.0134 0.21 0.95 4.4  79.5  191 4000
#>  [6]  19.6  149.5 0.00357 0.0083 0.19 0.96 4.9  99.4  325 4000
#>  [7]  30.1  266.6 0.00210 0.0062 0.16 0.89 5.2 160.3  496 4000
#>  [8]  48.7  449.3 0.00116 0.0041 0.14 0.91 6.0 254.9  939 4000
#>  [9]  83.3  900.9 0.00074 0.0027 0.12 0.90 6.7 401.5 1066 4000
#> [10] 147.8 1814.9 0.00056 0.0019 0.12 0.89 8.1 491.6 1916 4000