Interval plot

Create a plot based on a data frame providing endpoints of intervals, colors, line weights etc.

ivplot(
  X,
  name = "",
  file.name = "",
  split = NULL,
  Intervals = NULL,
  xlim,
  left.margin = 3,
  x.ticks = NULL,
  exp.labels = FALSE,
  xlab = "",
  title = "",
  top.axis = FALSE,
  use_color = TRUE,
  vline = NULL,
  device = "X11",
  size = c(297, 210)/25.4/2,
  font.family = "Courier",
  cex.label = NULL,
  ...
)

Arguments

X
A data frame providing data for creating one interval per row. See details below.

name
Name of file to produce

file.name
Name of file to produce

split
Name of column by which to divide the plot into groups.

Intervals
A list defining what intervals or dots to output per each row.

xlim
Numeric vector of length 2. Limits for the horizontal axis.

left.margin
Scalar. Size of left margin. If labels take too much space, increase this (default is 3)

x.ticks
Numeric vector.

exp.labels
Logical. Use log scale? Then print numeric values at x-ticks in the original (exponentiated) scale

xlab
Character.

title
Character; title.

top.axis
Logical. Print top axis?

use_color
Logical. Use color in plot or black?

vline
Scalar. Plot vertical line (will be plotted before intervals are

device
Character. To which device to output?

size
Numeric vector of length 2. Size of plot: vertical and horizontal sizes in inches.

font.family
Character. Font family (sans (Helvetica), serif (Times), mono (Courier), ...)

cex.label
number, a factor to shrink the 'cex' of the labels, between 0 and 1

...
Other arguments passed to plot

Value

The file name that was output; as a side effect a plot (a pdf file if device="pdf".)

Details

...