Distribution of the Sample Median

Compute the distribution sample median of the vector of values given as its argument.

# S3 method for rv
median(x, na.rm = FALSE, ...)

Arguments

x
a randomv vector containing the components whose distribution of the median value is to be computed.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

...
further arguments passed to median

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

See also

Examples

  x <- rvnorm(10) ## A random vector of length 10.
  median(x)       ## A random scalar (vector of length 1).

#>      mean   sd    1%  2.5%   25%   50%  75% 97.5%  99% sims
#> [1] 0.028 0.36 -0.82 -0.68 -0.19 0.026 0.25  0.71 0.79  200

  rvmedian(x)     ## A numeric vector of length 10.

#>  [1]  0.11054852 -0.03041283  0.08463253  0.02488842  0.06999428  0.16352824
#>  [7] -0.11388434 -0.03805603  0.04682688 -0.12771236