Distribution of the Range of a Random Vector

range.rv returns a 2-component random vector containing the distributions of the minimum and the maximum values of all the given arguments.

# S3 method for rv
range(..., na.rm = FALSE, finite = FALSE)

Arguments

...

further arguments passed to or from other methods

na.rm

logical, indicating if NAs should be omitted

finite

logical, indicating if all non-finite elements should be omitted

Details

This is the rv-compatible version of the function range.

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

See also

Author

Jouni Kerman jouni@kerman.com

Examples

  x <- rvnorm(mean=1:10, sd=1)
  print(range(x))
#>     name mean   sd   1% 2.5%  25%   50%  75% 97.5%  99% sims
#> [1]  min  0.8 0.90 -1.5 -1.0 0.19  0.81  1.4   2.5  2.8  200
#> [2]  max 10.2 0.81  8.6  8.8 9.57 10.12 10.7  11.8 11.9  200
  print(quantile(x, c(0,1)))
#>     name mean   sd   1% 2.5%  25%   50%  75% 97.5%  99% sims
#> [1]   0%  0.8 0.90 -1.5 -1.0 0.19  0.81  1.4   2.5  2.8  200
#> [2] 100% 10.2 0.81  8.6  8.8 9.57 10.12 10.7  11.8 11.9  200