Matrices and Arrays of Random Vectors

Arrange a given random vector into a matrix or array form.

rvmatrix(data = NA, nrow = 1, ncol = 1, byrow = FALSE,
  dimnames = NULL)

rvarray(data = NA, dim = length(data), dimnames = NULL)

Arguments

data
an optional data vector.

nrow
the desired number of rows.

ncol
the desired number of columns.

byrow
logical. If FALSE (the default) the matrix is filled by columns, otherwise the matrix is filled by rows.

dimnames
A dimnames attribute for the matrix: a list of length 2 giving the row and column names respectively.

dim
the dim attribute for the array to be created, that is a vector of length one or more giving the maximal indices in each dimension.

…
arguments passed to other methods

Details

These are 'rv' compatible versions of the functions matrix and array.

The function rvmatrix generates the random variable matrix via an rvarray call.

The rvarray function calls first array to set the dimensions of the argument data and then coerces the resulting array object to an 'rv' object.

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

See also

Examples

  n.rows <- 3; n.cols <- 4; n <- (n.rows*n.cols)
  mu.true <- rnorm(1:n.rows, mean=1:n.rows, sd=1)
  theta <- rvmatrix(rvnorm(n=n.cols, mean=mu.true, sd=0.5), nrow=n.rows)
  col.labels <- paste("Time", 1:n.cols, sep=":")
  row.labels <- paste("Unit", 1:n.rows, sep=":")
  dimnames(theta) <- list(row.labels, col.labels)
  print(theta)

#>          row    col   mean   sd   1% 2.5%  25% 50% 75% 97.5% 99% sims
#> [1,1] Unit:1 Time:1 :  1.3 0.50 0.19 0.41 1.00 1.3 1.6   2.3 2.6  200
#> [2,1] Unit:2 Time:1 :  1.8 0.56 0.69 0.79 1.46 1.8 2.2   2.9 3.0  200
#> [3,1] Unit:3 Time:1 :  2.8 0.49 1.75 1.93 2.48 2.9 3.2   3.8 3.9  200
#> [1,2] Unit:1 Time:2 :  1.4 0.53 0.34 0.38 1.03 1.3 1.7   2.4 2.6  200
#> [2,2] Unit:2 Time:2 :  1.8 0.48 0.66 0.88 1.47 1.8 2.1   2.7 2.8  200
#> [3,2] Unit:3 Time:2 :  2.9 0.54 1.58 1.89 2.49 2.9 3.3   3.9 4.0  200
#> [1,3] Unit:1 Time:3 :  1.4 0.47 0.44 0.53 0.96 1.3 1.6   2.3 2.4  200
#> [2,3] Unit:2 Time:3 :  1.7 0.50 0.75 0.79 1.37 1.7 2.1   2.7 2.8  200
#> [3,3] Unit:3 Time:3 :  2.9 0.51 1.79 1.89 2.56 2.9 3.3   4.0 4.1  200
#> [1,4] Unit:1 Time:4 :  1.4 0.51 0.20 0.33 1.02 1.4 1.7   2.3 2.4  200
#> [2,4] Unit:2 Time:4 :  1.8 0.52 0.56 0.81 1.47 1.8 2.1   2.7 3.0  200
#> [3,4] Unit:3 Time:4 :  2.9 0.52 1.59 1.91 2.50 2.9 3.2   3.9 3.9  200

  print(E(theta))

#>          Time:1   Time:2   Time:3   Time:4
#> Unit:1 1.327134 1.372999 1.350508 1.355914
#> Unit:2 1.810740 1.773406 1.712644 1.778697
#> Unit:3 2.838278 2.876188 2.917176 2.888863