Generate Random Vectors from a Beta Sampling Model

rvbeta generates a random vector from the beta sampling model;

rvbeta(n = 1, shape1, shape2)

Arguments

n

integer, number of random variables to generate

shape1

positive number or rv, 1st shape parameter

shape2

positive number or rv, 2nd shape parameter

Details

rvnbeta(n, a, b) ("neutral" Beta distribution) is equivalent to rvbeta(n, 1/3+a, 1/3+b).

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Author

Jouni Kerman jouni@kerman.com

Examples

   n <- 12         # sample size
   y <- (0:(n-1))  # observations
   a <- b <- 1/3   # the neutral beta prior
   rvbeta(1, shape1=a+y, shape2=b+n-y)
#>       mean    sd      1%    2.5%    25%   50%  75% 97.5%  99% sims
#>  [1] 0.029 0.046 1.3e-08 2.3e-06 0.0012 0.012 0.04  0.15 0.24  200
#>  [2] 0.113 0.093 8.8e-03 1.1e-02 0.0444 0.090 0.15  0.35 0.39  200
#>  [3] 0.192 0.102 2.6e-02 3.8e-02 0.1145 0.175 0.27  0.41 0.46  200
#>  [4] 0.270 0.123 6.7e-02 8.1e-02 0.1795 0.255 0.34  0.57 0.60  200
#>  [5] 0.347 0.126 1.0e-01 1.2e-01 0.2499 0.342 0.45  0.58 0.61  200
#>  [6] 0.420 0.134 1.2e-01 1.7e-01 0.3190 0.413 0.52  0.66 0.73  200
#>  [7] 0.508 0.136 2.4e-01 2.8e-01 0.3960 0.511 0.61  0.78 0.82  200
#>  [8] 0.577 0.133 2.8e-01 3.2e-01 0.4890 0.588 0.68  0.81 0.85  200
#>  [9] 0.670 0.132 3.4e-01 3.9e-01 0.5921 0.678 0.76  0.89 0.92  200
#> [10] 0.749 0.114 4.9e-01 5.2e-01 0.6838 0.766 0.83  0.92 0.95  200
#> [11] 0.809 0.103 5.6e-01 5.7e-01 0.7363 0.816 0.89  0.96 0.97  200
#> [12] 0.905 0.069 7.0e-01 7.5e-01 0.8565 0.917 0.96  1.00 1.00  200
   rvnbeta(1, shape1=y, shape2=n-y)
#>       mean    sd      1%    2.5%    25%    50%   75% 97.5%  99% sims
#>  [1] 0.028 0.046 9.9e-08 1.6e-06 0.0008 0.0074 0.036  0.15 0.24  200
#>  [2] 0.099 0.076 5.7e-03 1.0e-02 0.0419 0.0808 0.136  0.30 0.33  200
#>  [3] 0.190 0.108 2.4e-02 2.9e-02 0.1199 0.1760 0.255  0.46 0.47  200
#>  [4] 0.270 0.125 7.0e-02 8.5e-02 0.1785 0.2544 0.347  0.57 0.60  200
#>  [5] 0.328 0.128 8.6e-02 9.9e-02 0.2324 0.3153 0.411  0.56 0.59  200
#>  [6] 0.415 0.127 1.6e-01 1.9e-01 0.3278 0.4036 0.506  0.66 0.68  200
#>  [7] 0.495 0.152 1.7e-01 1.9e-01 0.3705 0.5047 0.605  0.78 0.81  200
#>  [8] 0.581 0.123 2.7e-01 3.4e-01 0.5080 0.5873 0.663  0.79 0.83  200
#>  [9] 0.645 0.129 3.7e-01 4.0e-01 0.5527 0.6534 0.735  0.89 0.92  200
#> [10] 0.740 0.106 4.6e-01 4.9e-01 0.6782 0.7574 0.810  0.91 0.93  200
#> [11] 0.811 0.113 4.6e-01 5.4e-01 0.7480 0.8221 0.896  0.97 0.98  200
#> [12] 0.892 0.089 6.2e-01 6.8e-01 0.8436 0.9131 0.962  0.99 1.00  200