Expectation of a Random Variable

rvmean

rvmean(x)

Arguments

x
an rv object

Value

A numerical vector with the same dimension as x.

Details

rvmean computes the means of the simulations of all individual components of a random vector (rv) object.

E is an alias for rvmean, standing for "Expectation."

Pr is another alias for rvmean, standing for "Probability of"; suggested to be used when the argument is a logical statement involving random variables (that is, a description of an event such as x>0 or x>y). Then Pr(x>0) gives the probability of the event "x>0". The statement x>0 returns a Bernoulli (indicator) random variable object (having 1/0 or TRUE/FALSE values) and the expectation of such variable is just the probability of the event where the indicator is one.

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

See also

Examples

  x <- rvnorm(mean=(1:10)/5, sd=1)
  rvmean(x)  # means of the 10 components

#>  [1] 0.2050028 0.2978910 0.7058464 0.8110282 1.1072214 1.3309428 1.3338731
#>  [8] 1.5973089 1.8595260 1.9999273

  E(x)       # same as rvmean(x)

#>  [1] 0.2050028 0.2978910 0.7058464 0.8110282 1.1072214 1.3309428 1.3338731
#>  [8] 1.5973089 1.8595260 1.9999273

  Pr(x>1)    # probabilities that each component is >1.

#>  [1] 0.230 0.265 0.380 0.400 0.555 0.675 0.645 0.735 0.805 0.865