Generate Random Variables from a Gaussian (Normal) Sampling Model

Generates a random vector from a Gaussian sampling model.

rvnorm(n = 1, mean = 0, sd = 1, var = NULL, precision)

Arguments

n
integer: number of variables to generate.

mean
mean, may be a rv

sd
standard deviation; scalar or vector (constant or rv, not matrix)

var
variance, can be given instead of sd. Scalar, vector, or matrix.

precision
inverse variance or variance matrix, may be given instead of sd or var

Value

An rv object of length n times the length of the mean vector.

If mean is a vector, a vector is returned: n refers to how many vectors or scalars are replicated.

Note

If any of the arguments are random, the resulting simulations may have non-normal marginal distributions; for example, if an inverse-chi-squared scalar rv var and zero mean is given, the resulting rv will have a t-distribution.

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Examples

  x <- rvnorm(mean=1:10, sd=1:10) # A vector of length 10.
  Sigma <- diag(1:10)
  y <- rvnorm(mean=1:10, var=Sigma)