Set or Query Parameters of the 'rv' Package

Sets or retrieves parameters of the rv package.

rvpar(...)

Arguments

...

arguments in tag = value form, or a list or character vector of tagged values. The available tags are described below.

Value

In the case of a single tag query, the requested value. In the case of multiple tag query, a list of requested values.

Details

rvcol

color of a random point (interval), such as 'red' or 'blue'

rvlex

middle interval expansion factor

rvlwd

line weight of a random interval

print.digits

number of digits to show in the summaries

rvpoint

what to output when plotting a random point; default list("95%", "50%", "mean")

point.sample

number of points to plot when plotting a rv-rv scatterplot. Default 400.

line.sample

number of lines to draw when plotting a random sample of lines (see abline). Default 20.

summary.dimnames

logical; output dimnames in the summary of an rv object? Default TRUE.

summary.quantiles.numeric

vector of quantiles to compute for the summary of a numeric rv object.

summary.quantiles.integer

vector of quantiles to compute for the summary of an integer-valued rv object. By default contains 0 and 1 (for the min and max values).

References

Kerman, J. and Gelman, A. (2007). Manipulating and Summarizing Posterior Simulations Using Random Variable Objects. Statistics and Computing 17:3, 235-244.

See also vignette("rv").

Author

Jouni Kerman jouni@kerman.com

Examples

  rvpar()$rvcol
#> [1] "default"
  rvpar("rvcol")
#> [1] "default"